+++ .vscode/tasks.json
@@ -2,7 +2,7 @@
 	"version": "2.0.0",
 	"tasks": [
 		{
-			"label": "cnrNpmStart",
+			"label": "cnrNodeStartWithDebugger",
 			"type": "npm",
 			"script": "start",
 			"problemMatcher": [],
