+++ views/pages/activities.ejs
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <% include ../includes/header.ejs %>
+  </head>
+  <body>
+    <% include ../includes/nav.ejs %>
+      <div class="jumbotron text-center">
+        <div class="container">
+          <h1>
+            <%= title %>
+          </h1>
+          <p>Team 2 - Activity 01</p>
+        </div>
+      </div>
+      <!-- add activity form -->
+      <div class="container">
+        <form method="POST" action="/ta01/add-activity">
+          <input type="text" name="newActivity" placeholder="Activity">
+          <input type="submit" value="Add Activity">
+        </form>
+        <!-- delete activity form -->
+        <form method="POST" action="/ta01/delete-activity">
+          <input type="text" name="deleteActivity" placeholder="Delete Activity">
+          <input type="submit" value="Delete Activity">
+        </form>
+        <!-- display activities -->
+        <% if(activities.length > 0){ %>
+          <% for(i=0; i<numActivities; i++) {%>
+            <div></div>
+            <%= activities[i]%>
+          <%}} %>
+      </div>
+  </body>
+</html>
