+++ views/pages/ta01.ejs
@@ -1,0 +1,27 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <% include ../includes/header.ejs %>
+  </head>
+  <body>
+    <% include ../includes/nav.ejs %>
+      <div class="jumbotron text-center">
+        <div class="container">
+          <h1>
+            <%= title %>
+          </h1>
+          <p>Team 2 - Activity 01</p>
+        </div>
+      </div>
+      
+      <!-- content -->
+      <div style="margin-left:3rem;">
+        <h1>Welcome to my world!</h1>
+        <a href="/">Home</a></br>
+        <a href="ta01/activities">Activities List</a></br>
+        <a href="ta01/stretch-1">Stretch 1 (CSS)</a></br>
+        <a href="ta01/stretch-2">Stretch 2 (Write Form input to text input)</a></br>
+        <a href="ta01/stretch-3">Stretch 3 (Add two number inputs together)</a></br>
+      </div>
+  </body>
+</html>
