+++ views/pages/ta02.ejs
@@ -1,38 +1,35 @@
 <!DOCTYPE html>
 <html>
-
-<head>
-  <% include ../includes/header.ejs %>
-</head>
-
-<body>
-  <% include ../includes/nav.ejs %>
-    <div class="jumbotron text-center">
-      <div class="container">
-        <h1>
-          <%= title %>
-        </h1>
-        <p>This is the placeholder for Team Activity 02</p>
+  <head>
+    <% include ../includes/header.ejs %>
+  </head>
+  <body>
+    <% include ../includes/nav.ejs %>
+      <div class="jumbotron text-center">
+        <div class="container">
+          <h1>
+            <%= title %>
+          </h1>
+          <p>Team 2 - Activity 02</p>
+        </div>
       </div>
-    </div>
-    <div class="container">
-      <form method="POST" action="/ta02/add-user">
-        <input type="text" placeholder="Username" name="username">
-        <input type="submit" value="Add User">
-      </form>
-      <% if(users.length > 0){ %>
-        <% for(i=0; i<numUsers; i++) {%>
-          <div></div>
-          <%= users[i]%>
-        <%}} %>
-
-        <form method="POST" action="/ta02/delete-user">
-          <input type="text" name="deleteName" placeholder="Delete Name">
+      <!-- add username form -->
+      <div class="container">
+        <form method="POST" action="/ta02/add-username">
+          <input type="text" name="newUserName" placeholder="username">
+          <input type="submit" value="Add User">
+        </form>
+        <!-- delete username form -->
+        <form method="POST" action="/ta02/delete-username">
+          <input type="text" name="deleteUserName" placeholder="Delete Name">
           <input type="submit" value="Delete">
         </form>
-    </div>
-
-
-</body>
-
+        <!-- display users -->
+        <% if(users.length > 0){ %>
+          <% for(i=0; i<numUsers; i++) {%>
+            <div></div>
+            <%= users[i]%>
+          <%}} %>
+      </div>
+  </body>
 </html>
